Split treeworksConsent into two separate types
Despite the [standard form][1] combining both types of tree works
they follow different workflows - trees with a Tree Preservation
Order (TPO) is an [application for consent][2] whereas trees in a
Conservation Area (CA) is a [notification of intended works][3].

The process for trees with a TPO takes eight weeks and the process
for trees in a CA is six weeks. Trees in a conservation area with
an existing TPO follow the standard procedure.

[1]: https://assets.publishing.service.gov.uk/media/5b06c81de5274a1c5490ce6f/Form031_england_en_revised__2_.pdf
[2]: https://www.gov.uk/guidance/tree-preservation-orders-and-trees-in-conservation-areas#making-applications-tpo
[3]: https://www.gov.uk/guidance/tree-preservation-orders-and-trees-in-conservation-areas#Protecting-trees-in-conservation-areas